Round timeline average to the nearest integer instead of truncating decimals

+++ src/main/java/pl/llp/aircasting/screens/stream/MeasurementPresenter.java
@@ -311,7 +311,7 @@ public class MeasurementPresenter implements SharedPreferences.OnSharedPreferenc
         }
 
         mTimelinePeak = peakValue;
-        mTimelineAverage = sum / timelineMeasurements.size();
+        mTimelineAverage = sum / (double) timelineMeasurements.size();
     }
 
     public void registerListener(Listener listener) {
